# Improve perf of `ActionView::Helpers::TextHelper#excerpt` for large strings.