@eugeneius 2020-10-13
The pull request title and commit message both mention `truncate`, when the optimisation is actually for `excerpt`.