@tgxworld 2020-12-23
@eugeneius Sorry for the late reply as I got caught up with a project at work. I've updated the PR and commit as per your review.